webapp/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pdfplumber>=0.11.0
python-dotenv>=1.0.0
litellm>=1.64.0
dspy-ai>=2.6.27
stix2>=3.0.0
structlog>=24.0.0
orjson>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
